Use DI for RandomAccessCallback, move Wrapper to outer class
